+++ binding-hs/CHANGELOG.md
@@ -1,3 +1,163 @@
+# Version 0.1.0.2
+
+[tag v0.1.0.2](https://github.com/stites/gym-http-api/releases/tag/v0.1.0.2)
+
+- Expose all environments listed at https://gym.openai.com/envs/ as of 07/17/2018. Current environments supported includes:
+  + Classic Control Environments:
+    + CartPoleV0
+    + CartPoleV1
+    + AcrobotV1
+    + MountainCarV0
+    + MountainCarContinuousV0
+    + PendulumV0
+  + Toy Text Environments
+    + BlackjackV0
+    + FrozenLakeV0
+    + FrozenLake8x8V0
+    + GuessingGameV0
+    + HotterColderV0
+    + NChainV0
+    + RouletteV0
+    + TaxiV2
+  + Box2D Environments:
+    + BipedalWalkerV2
+    + BipedalWalkerHardcoreV2
+    + CarRacingV0
+    + LunarLanderV2
+    + LunarLanderContinuousV2
+  + Algorithm Environments:
+    + CopyV0
+    + DuplicatedInputV0
+    + RepeatCopyV0
+    + ReverseV0
+    + ReversedAdditionV0
+    + ReversedAddition3V0
+  + Atari Environments:
+    + AirRaidRamV0
+    + AirRaidV0
+    + AlienRamV0
+    + AlienV0
+    + AmidarRamV0
+    + AmidarV0
+    + AssaultRamV0
+    + AssaultV0
+    + AsterixRamV0
+    + AsterixV0
+    + AsteroidsRamV0
+    + AsteroidsV0
+    + AtlantisRamV0
+    + AtlantisV0
+    + BankHeistRamV0
+    + BankHeistV0
+    + BattleZoneRamV0
+    + BattleZoneV0
+    + BeamRiderRamV0
+    + BeamRiderV0
+    + BerzerkRamV0
+    + BerzerkV0
+    + BowlingRamV0
+    + BowlingV0
+    + BoxingRamV0
+    + BoxingV0
+    + BreakoutRamV0
+    + BreakoutV0
+    + CarnivalRamV0
+    + CarnivalV0
+    + CentipedeRamV0
+    + CentipedeV0
+    + ChopperCommandRamV0
+    + ChopperCommandV0
+    + CrazyClimberRamV0
+    + CrazyClimberV0
+    + DemonAttackRamV0
+    + DemonAttackV0
+    + DoubleDunkRamV0
+    + DoubleDunkV0
+    + ElevatorActionRamV0
+    + ElevatorActionV0
+    + EnduroRamV0
+    + EnduroV0
+    + FishingDerbyRamV0
+    + FishingDerbyV0
+    + FreewayRamV0
+    + FreewayV0
+    + FrostbiteRamV0
+    + FrostbiteV0
+    + GopherRamV0
+    + GopherV0
+    + GravitarRamV0
+    + GravitarV0
+    + IceHockeyRamV0
+    + IceHockeyV0
+    + JamesbondRamV0
+    + JamesbondV0
+    + JourneyEscapeRamV0
+    + JourneyEscapeV0
+    + KangarooRamV0
+    + KangarooV0
+    + KrullRamV0
+    + KrullV0
+    + KungFuMasterRamV0
+    + KungFuMasterV0
+    + MontezumaRevengeRamV0
+    + MontezumaRevengeV0
+    + MsPacmanRamV0
+    + MsPacmanV0
+    + NameThisGameRamV0
+    + NameThisGameV0
+    + PhoenixRamV0
+    + PhoenixV0
+    + PitfallRamV0
+    + PitfallV0
+    + PongRamV0
+    + PongV0
+    + PooyanRamV0
+    + PooyanV0
+    + PrivateEyeRamV0
+    + PrivateEyeV0
+    + QbertRamV0
+    + QbertV0
+    + RiverraidRamV0
+    + RiverraidV0
+    + RoadRunnerRamV0
+    + RoadRunnerV0
+    + RobotankRamV0
+    + RobotankV0
+    + SeaquestRamV0
+    + SeaquestV0
+    + SkiingRamV0
+    + SkiingV0
+    + SolarisRamV0
+    + SolarisV0
+    + SpaceInvadersRamV0
+    + SpaceInvadersV0
+    + StarGunnerRamV0
+    + StarGunnerV0
+    + TennisRamV0
+    + TennisV0
+    + TimePilotRamV0
+    + TimePilotV0
+    + TutankhamRamV0
+    + TutankhamV0
+    + UpNDownRamV0
+    + UpNDownV0
+    + VentureRamV0
+    + VentureV0
+    + VideoPinballRamV0
+    + VideoPinballV0
+    + WizardOfWorRamV0
+    + WizardOfWorV0
+    + YarsRevengeRamV0
+    + YarsRevengeV0
+    + ZaxxonRamV0
+    + ZaxxonV0
+
+# Version 0.1.0.1
+
+[tag v0.1.0.1](https://github.com/stites/gym-http-api/releases/tag/v0.1.0.1)
+
+- Bump compatability with lts-11
+
 # Version 0.1.0.0
 
 [tag v0.1.0.0](https://github.com/stites/gym-http-api/releases/tag/v0.1.0.0)